requirements.txt
fastapi
numpy
Pillow
opencv-python
tensorflow==2.18.0
pandas
scikit-learn
matplotlib
python-multipart
uvicorn
